+++ tests/test_adjustment.py
@@ -24,6 +24,7 @@ from xsdba.adjustment import (
 from xsdba.base import Grouper, stack_periods
 from xsdba.options import set_options
 from xsdba.processing import (
+    adapt_freq,
     jitter_over_thresh,
     jitter_under_thresh,
     stack_variables,
@@ -419,6 +420,29 @@ class TestDQM:
             nquantiles=50,
         )
         assert DQM.ds.sizes == {"dayofyear": 360, "quantiles": 50}
+
+    @pytest.mark.parametrize("group", ["time", "time.month"])
+    def test_adapt_freq_grouping(self, cannon_2015_rvs, random, group):
+        ref, hist, sim = cannon_2015_rvs(15000, random=random)
+
+        dqm = DetrendedQuantileMapping.train(
+            ref, hist, kind="*", group=group, adapt_freq_thresh="1 kg m-2 d-1"
+        )
+        dqm.adjust(sim)
+
+    def test_adapt_freq_time_explicit(self, cannon_2015_rvs, random):
+        ref, hist, _ = cannon_2015_rvs(15000, random=random)
+        thr = "1 kg m-2/d"
+        ref = jitter_under_thresh(ref, "0.1   kg m-2 / d")
+        hist = jitter_under_thresh(hist, "0.1 kg m-2 / d")
+        hist_ad, _, _ = adapt_freq(ref, hist, group="time", thresh=thr)
+        ADJ = DetrendedQuantileMapping.train(
+            ref, hist, kind="*", group="time", adapt_freq_thresh=thr
+        )
+        out = ADJ.adjust(hist)
+        ADJ.adapt_freq_thresh = None
+        out_ad = ADJ.adjust(hist_ad)
+        np.testing.assert_allclose(out.values, out_ad.values)
 
 
 @pytest.mark.slow
@@ -984,6 +1008,25 @@ class TestExtremeValues:
         assert (scen2.where(exval) > EX.ds.thresh).sum() > (
             scen.where(exval) > EX.ds.thresh
         ).sum()
+
+    def test_quantified_cluster_thresh(self, gosset):
+        dsim = xr.open_dataset(gosset.fetch("sdba/CanESM2_1950-2100.nc"))  # .chunk()
+        dref = xr.open_dataset(gosset.fetch("sdba/ahccd_1950-2013.nc"))  # .chunk()
+        ref = dref.sel(time=slice("1950", "2009")).pr
+        hist = dsim.sel(time=slice("1950", "2009")).pr
+        # TODO: Do we want to include standard conversions in xsdba tests?
+        # this is just convenient for now to keep those tests
+        hist = pint_multiply(hist, "1e-03 m^3/kg")
+        hist = convert_units_to(hist, ref)
+
+        EX = ExtremeValues.train(ref, hist, cluster_thresh="1 mm/day", q_thresh=0.97)
+        scen = EX.adjust(hist, hist, frac=0.000000001)
+        cluster_thresh = xr.DataArray(1, attrs={"units": "mm/d"})
+        EXQ = ExtremeValues.train(
+            ref, hist, cluster_thresh=cluster_thresh, q_thresh=0.97
+        )
+        scenQ = EXQ.adjust(hist, hist, frac=0.000000001)
+        assert (scen.values == scenQ.values).all()
 
     @pytest.mark.slow
     def test_real_data(self, gosset):