+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ Changes
 * Throw an error if `group=Grouper('5D',window)` is used with a biasadjust method other than `MBCn`.
 * ``xsdba.processing.to_additive_space`` accepts `clip_next_to_bounds`, which avoids infinities by ensuring `lower_bound < data < upper_bound`. (:issue:`164`, :pull:`165`).
 * Allow nan values in ``xsdba.measures.rmse`` and ``xsdba.measures.mae``. (:pull:`170`).
+* The adaptation of frequencies through `adapt_freq_thresh_value` is now applied in the adjusting step as well. (:pull:`160`).
+* ``xsdba.adjustment.ExtremeValues`` now accepts a DataArray for `cluster_thresh`, letting specify distinct thresholds for multiple locations. (:issue:`177`, :pull:`179`).
 
 Fixes
 ^^^^^
@@ -32,6 +34,7 @@ Internal changes
 ^^^^^^^^^^^^^^^^
 * The `tox` and CI configurations now support the installation of `SBCK` and `Eigen3` for testing purposes. (:pull:`139`).
 * The `coveralls` tox keyword has been renamed to `coverage` to avoid confusion with the `coveralls` service. (:pull:`139`).
+* The order of arguments in the following private functions was changed: ``xsdba._adjustment.{_fit_on_cluster,_fit_cluster_and_cdf, _extremes_train_1d}``.
 
 .. _changes_0.4.0: